+++ pkg_testing_tools/main.py
@@ -124,8 +124,9 @@ def process_args(sysargs):
         type=str,
         required=False,
         default="once",
-        choices=["once", "always", "force", "never"],
-        help="Enables FEATURES='test' once, for default use flags, always, for every run or never. force also sets EBUILD_FORCE_TEST=1. Default: 'once'.",
+        choices=["once", "always", "force", "never", "first"],
+        # always respects if the package has tests or not to set FEATURES='test', whereas force always sets it, even if package has no tests.
+        help="Enables FEATURES='test' once, for default use flags, always, for every run or never. force also sets EBUILD_FORCE_TEST=1. Default: 'once'. Meaning tests are executed last.",
     )
 
     optional.add_argument(