+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `--depclean/-c` flag: to perform 'emerge --depclean' before each test install
+- Add `--test-feature-scope=first` to run test feature toggle first
+
+### Changed
+
+- Revert to running tests lasts unless new `--test-feature-scope=first` is set
+
 
 ## [0.2.7] - 2025-11-24